functions/pubsub-worker/requirements.txt
Flask
gunicorn
psycopg2-binary
requests
beautifulsoup4